Supporting recursive fetch
We now do not issue a reset on connect. A client-side fetch is required.

Client side now saves socket so it can be called on a fetch. Any provided filter would be inside options object, and sent along with the fetch to the server side. Note: on the server side, fetch is not overridden - it calls straight through to sync, which hits red is.

TODO: Finish the work to filter when sending data down. Potentially have different behavior for different kinds of events (add vs. reset).